lingfield_project/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.0.8
django-crispy-forms==1.9.2
filelock==3.0.12
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.28
wrapt==1.12.1
